+++ drivers/idle/wult/main.c
@@ -407,13 +407,13 @@ void wult_unregister(void)
 }
 EXPORT_SYMBOL_GPL(wult_unregister);
 
-static const struct x86_cpu_id intel_cpu_ids[] = {
-	X86_MATCH_VENDOR_FAM_FEATURE(INTEL, 6, X86_FEATURE_CONSTANT_TSC, NULL),
-	X86_MATCH_VENDOR_FAM_FEATURE(AMD, 23, X86_FEATURE_CONSTANT_TSC, NULL),
-	X86_MATCH_VENDOR_FAM_FEATURE(AMD, 25, X86_FEATURE_CONSTANT_TSC, NULL),
+static const struct x86_cpu_id stable_tsc_cpu_ids[] = {
+	X86_MATCH_VENDOR_FAM_FEATURE(INTEL, X86_FAMILY_ANY, X86_FEATURE_CONSTANT_TSC, NULL),
+	X86_MATCH_VENDOR_FAM_FEATURE(AMD, X86_FAMILY_ANY, X86_FEATURE_CONSTANT_TSC, NULL),
+	X86_MATCH_VENDOR_FAM_FEATURE(AMD, X86_FAMILY_ANY, X86_FEATURE_CONSTANT_TSC, NULL),
 	{}
 };
-MODULE_DEVICE_TABLE(x86cpu, intel_cpu_ids);
+MODULE_DEVICE_TABLE(x86cpu, stable_tsc_cpu_ids);
 
 /* Module initialization function. */
 static int __init wult_init(void)
@@ -425,9 +425,9 @@ static int __init wult_init(void)
 		return -EINVAL;
 	}
 
-	id = x86_match_cpu(intel_cpu_ids);
+	id = x86_match_cpu(stable_tsc_cpu_ids);
 	if (!id) {
-		wult_err("unsupported CPU family: an x86 CPU with constant TSC is required");
+		wult_err("unsupported CPU: an x86 CPU with constant TSC is required");
 		return -EINVAL;
 	}